Appease staticlint in func_test
