+++ .github/workflows/test.yml
@@ -24,6 +24,7 @@ jobs:
       - name: Lint
         run: |
           staticcheck ./...
+          staticcheck -tags=func_test ./...
           go vet ./...
           go vet -tags=func_test ./...
